Decompose dm-tool BookBrowser.tsx into 7 sub-modules
First slice of the C-e BookBrowser/BookReader decomposition. Strict
mechanical extraction — zero behaviour change, every type / helper /
component moves verbatim with imports adjusted.

`apps/dm-tool/src/features/book-browser/BookBrowser.tsx` was 1128 LOC.
After this slice it's ~470 LOC. Extracted pieces live under a new
`BookBrowser/` directory:

- `BookBrowser/types.ts` — `CatalogEntry`, `TabDef`, `PersistedTab`
- `BookBrowser/constants.ts` — `BROWSER_TAB`, `SYSTEM_ORDER`,
  `CATEGORY_ORDER`, plus the `CARD_WIDTH` / `CARD_HEIGHT` / `GAP` /
  `SECTION_HEIGHT` layout values shared by `cards.tsx` and
  `CatalogGrid.tsx`
- `BookBrowser/helpers.ts` — `persistTabs`, `loadPersistedTabs`,
  `effectiveSystem` / `Category` / `Publisher` / `Title`,
  `isApCategory`
- `BookBrowser/sidebar.tsx` — `NavItem`, `NavGroup`, `SystemGroup`
  (the category rail trio)
- `BookBrowser/CatalogGrid.tsx` — virtualised catalog grid + the
  `LayoutRow` type it uses internally
- `BookBrowser/cards.tsx` — `BookCard`, `ApCard`, `CoverArea`,
  `HoverMeta`, `SystemBadge`, `RulesetBadge`
- `BookBrowser/BookContextMenu.tsx` — context menu + its
  `CTX_CATEGORIES` / `CTX_SYSTEMS` lists

`BookBrowser.tsx` (the route entry) keeps the root component plus
`TabBar` (small, only used by main).

`App.tsx` import (`from './features/book-browser/BookBrowser'`) keeps
resolving to the .tsx file — TypeScript's bundler resolution prefers
the file over the sibling directory.

Bucket C-e1 of the cleanup audit. C-e2 (BookReader.tsx, 1154 LOC)
follows in its own PR.
