+++ apps/dm-tool/src/features/book-browser/BookBrowser.tsx
@@ -1,83 +1,30 @@
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
-import { useVirtualizer } from '@tanstack/react-virtual';
-import { Check, ChevronRight, Library, Layers, RefreshCw, Sparkles, X } from 'lucide-react';
+import { useCallback, useEffect, useMemo, useState } from 'react';
+import { Library, RefreshCw, Sparkles, X } from 'lucide-react';
 import { ResizableSidebar } from '@/components/ResizableSidebar';
 import { Button } from '@/components/ui/button';
 import { ScrollArea } from '@/components/ui/scroll-area';
 import { Separator } from '@/components/ui/separator';
 import { cn } from '@/lib/utils';
 import { api } from '@/lib/api';
-import { STORAGE_KEYS } from '@/lib/constants';
-import { readJson, readString, writeJson, writeString } from '@/lib/storage-utils';
+import type { Book } from '@foundry-toolkit/shared/types';
 import { useBackgroundIngest, useBookClassify, useBookList, useBookScan } from './useBooks';
 import { BookReader } from './BookReader';
-import { groupAdventurePaths, apTotalPages, type ApGroup } from './ap-merge';
-import type { Book } from '@foundry-toolkit/shared/types';
+import { groupAdventurePaths } from './ap-merge';
 
-// ---------------------------------------------------------------------------
-// Types
-// ---------------------------------------------------------------------------
-
-type CatalogEntry = { kind: 'book'; book: Book } | { kind: 'ap'; group: ApGroup } | { kind: 'section'; label: string };
-
-type TabDef =
-  | { id: 'browser'; kind: 'browser' }
-  | { id: string; kind: 'book'; bookId: number; title: string }
-  | { id: string; kind: 'ap'; group: ApGroup; title: string };
-
-const BROWSER_TAB: TabDef = { id: 'browser', kind: 'browser' };
-
-/** Serializable shape saved to localStorage. */
-type PersistedTab =
-  | { id: 'browser'; kind: 'browser' }
-  | { id: string; kind: 'book'; bookId: number; title: string }
-  | { id: string; kind: 'ap'; subcategory: string; title: string };
-
-function persistTabs(tabs: TabDef[], activeTabId: string) {
-  const serializable: PersistedTab[] = tabs.map((t) => {
-    if (t.kind === 'ap') return { id: t.id, kind: 'ap', subcategory: t.group.subcategory, title: t.title };
-    return t as PersistedTab;
-  });
-  writeJson(STORAGE_KEYS.bookTabs, serializable);
-  writeString(STORAGE_KEYS.bookActiveTab, activeTabId);
-}
-
-function loadPersistedTabs(): { tabs: PersistedTab[]; activeTabId: string } | null {
-  const tabs = readJson<PersistedTab[] | null>(STORAGE_KEYS.bookTabs, null);
-  if (!tabs) return null;
-  const active = readString(STORAGE_KEYS.bookActiveTab) ?? 'browser';
-  return { tabs, activeTabId: active };
-}
-
-// Use AI-derived classification when available, fall back to folder-derived.
-// For folder-derived books: category = system (e.g. "PF2e"), subcategory = category type (e.g. "Rulebooks").
-function effectiveSystem(b: Book): string {
-  return b.aiSystem ?? b.category;
-}
-function effectiveCategory(b: Book): string {
-  return b.aiCategory ?? b.subcategory ?? 'Uncategorized';
-}
-function effectivePublisher(b: Book): string | null {
-  return b.aiPublisher ?? null;
-}
-function effectiveTitle(b: Book): string {
-  return b.aiTitle ?? b.title;
-}
-
-const isApCategory = (n: string) => n === 'Adventure Path' || n === 'Adventure Paths';
-const SYSTEM_ORDER = ['PF2e', '5e', 'Generic'];
-const CATEGORY_ORDER = [
-  'Rulebook',
-  'Adventure Path',
-  'Adventure',
-  'Setting',
-  'Supplement',
-  'Rulebooks',
-  'Adventure Paths',
-  'Adventures',
-  'Lost Omens',
-  'Beginner Box',
-];
+import { BookContextMenu } from './BookBrowser/BookContextMenu';
+import { CatalogGrid } from './BookBrowser/CatalogGrid';
+import { BROWSER_TAB, CATEGORY_ORDER, SYSTEM_ORDER } from './BookBrowser/constants';
+import {
+  effectiveCategory,
+  effectivePublisher,
+  effectiveSystem,
+  effectiveTitle,
+  isApCategory,
+  loadPersistedTabs,
+  persistTabs,
+} from './BookBrowser/helpers';
+import { NavItem, SystemGroup } from './BookBrowser/sidebar';
+import type { CatalogEntry, TabDef } from './BookBrowser/types';
 
 // ---------------------------------------------------------------------------
 // Top-level component
@@ -540,589 +487,6 @@ function TabBar({
           </button>
         );
       })}
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Category rail components
-// ---------------------------------------------------------------------------
-
-/** Leaf nav item — no chevron, just a clickable label + count. */
-function NavItem({
-  name,
-  count,
-  active,
-  onClick,
-  indent = 0,
-}: {
-  name: string;
-  count: number;
-  active: boolean;
-  onClick: () => void;
-  indent?: number;
-}) {
-  return (
-    <button
-      type="button"
-      onClick={onClick}
-      className={cn(
-        'flex w-full items-center justify-between pr-3 py-1.5 text-left text-xs transition-colors',
-        active
-          ? 'bg-accent text-foreground font-medium'
-          : 'text-muted-foreground hover:bg-accent/50 hover:text-foreground',
-      )}
-      style={{ paddingLeft: 12 + indent * 12 }}
-    >
-      <span className="truncate">{name}</span>
-      <span className="ml-2 shrink-0 tabular-nums text-[10px] opacity-60">{count}</span>
-    </button>
-  );
-}
-
-/** Expandable nav group — chevron + label + count, with children. */
-function NavGroup({
-  name,
-  count,
-  active,
-  indent,
-  expanded,
-  onToggle,
-  onClick,
-  children,
-}: {
-  name: string;
-  count: number;
-  active: boolean;
-  indent: number;
-  expanded: boolean;
-  onToggle: () => void;
-  onClick: () => void;
-  children: React.ReactNode;
-}) {
-  return (
-    <div>
-      <div className="flex items-center" style={{ paddingLeft: indent * 12 }}>
-        <button
-          type="button"
-          className="flex h-6 w-5 items-center justify-center text-muted-foreground"
-          onClick={(e) => {
-            e.stopPropagation();
-            onToggle();
-          }}
-        >
-          <ChevronRight className={cn('h-3 w-3 transition-transform', expanded && 'rotate-90')} />
-        </button>
-        <button
-          type="button"
-          onClick={onClick}
-          className={cn(
-            'flex flex-1 items-center justify-between py-1.5 pr-3 text-left text-xs transition-colors',
-            active
-              ? 'bg-accent text-foreground font-medium'
-              : 'text-muted-foreground hover:bg-accent/50 hover:text-foreground',
-          )}
-        >
-          <span className="truncate">{name}</span>
-          <span className="ml-2 shrink-0 tabular-nums text-[10px] opacity-60">{count}</span>
-        </button>
-      </div>
-      {expanded && children}
-    </div>
-  );
-}
-
-/** Top-level system group with nested category → publisher tree. */
-function SystemGroup({
-  system,
-  selectedSystem,
-  selectedCategory,
-  selectedPublisher,
-  expandedKeys,
-  onToggle,
-  onSelect,
-}: {
-  system: {
-    name: string;
-    categories: { name: string; publishers: { name: string; count: number }[]; count: number }[];
-    count: number;
-  };
-  selectedSystem: string | null;
-  selectedCategory: string | null;
-  selectedPublisher: string | null;
-  expandedKeys: Set<string>;
-  onToggle: (key: string) => void;
-  onSelect: (sys: string | null, cat: string | null, pub: string | null) => void;
-}) {
-  const sysActive = selectedSystem === system.name && !selectedCategory;
-  return (
-    <NavGroup
-      name={system.name}
-      count={system.count}
-      active={sysActive}
-      indent={0}
-      expanded={expandedKeys.has(system.name)}
-      onToggle={() => onToggle(system.name)}
-      onClick={() => onSelect(sysActive ? null : system.name, null, null)}
-    >
-      {system.categories.map((cat) => {
-        const catActive = selectedSystem === system.name && selectedCategory === cat.name && !selectedPublisher;
-        const hasPubs =
-          cat.publishers.length > 1 || (cat.publishers.length === 1 && cat.publishers[0]!.name !== 'Unknown');
-        const catKey = `${system.name}/${cat.name}`;
-        return hasPubs ? (
-          <NavGroup
-            key={cat.name}
-            name={cat.name}
-            count={cat.count}
-            active={catActive}
-            indent={1}
-            expanded={expandedKeys.has(catKey)}
-            onToggle={() => onToggle(catKey)}
-            onClick={() => onSelect(system.name, catActive ? null : cat.name, null)}
-          >
-            {cat.publishers.map((pub) => (
-              <NavItem
-                key={pub.name}
-                name={pub.name}
-                count={pub.count}
-                indent={4}
-                active={
-                  selectedSystem === system.name && selectedCategory === cat.name && selectedPublisher === pub.name
-                }
-                onClick={() => {
-                  const pubActive =
-                    selectedSystem === system.name && selectedCategory === cat.name && selectedPublisher === pub.name;
-                  onSelect(system.name, cat.name, pubActive ? null : pub.name);
-                }}
-              />
-            ))}
-          </NavGroup>
-        ) : (
-          <NavItem
-            key={cat.name}
-            name={cat.name}
-            count={cat.count}
-            indent={3}
-            active={catActive}
-            onClick={() => onSelect(system.name, catActive ? null : cat.name, null)}
-          />
-        );
-      })}
-    </NavGroup>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Virtualized catalog grid — handles Book, ApGroup, and section headers
-// ---------------------------------------------------------------------------
-
-const CARD_WIDTH = 160;
-const CARD_HEIGHT = 240;
-const GAP = 12;
-const SECTION_HEIGHT = 36;
-
-/** A layout row is either a section header (full width) or a row of cards. */
-type LayoutRow = { kind: 'section'; label: string } | { kind: 'cards'; items: CatalogEntry[] };
-
-function CatalogGrid({
-  entries,
-  onSelect,
-  onBookContextMenu,
-}: {
-  entries: CatalogEntry[];
-  onSelect: (e: CatalogEntry) => void;
-  onBookContextMenu?: (e: React.MouseEvent, book: Book) => void;
-}) {
-  const parentRef = useRef<HTMLDivElement>(null);
-  const [columnCount, setColumnCount] = useState(1);
-
-  useEffect(() => {
-    const el = parentRef.current;
-    if (!el) return;
-    const update = () => {
-      const w = el.clientWidth;
-      setColumnCount(Math.max(1, Math.floor((w + GAP) / (CARD_WIDTH + GAP))));
-    };
-    update();
-    const ro = new ResizeObserver(update);
-    ro.observe(el);
-    return () => ro.disconnect();
-  }, []);
-
-  // Build layout rows: section headers get their own row, card entries
-  // are chunked into rows of `columnCount`.
-  const layoutRows = useMemo((): LayoutRow[] => {
-    const rows: LayoutRow[] = [];
-    const cardBuffer: CatalogEntry[] = [];
-
-    const flushCards = () => {
-      while (cardBuffer.length > 0) {
-        rows.push({ kind: 'cards', items: cardBuffer.splice(0, columnCount) });
-      }
-    };
-
-    for (const entry of entries) {
-      if (entry.kind === 'section') {
-        flushCards();
-        rows.push({ kind: 'section', label: entry.label });
-      } else {
-        cardBuffer.push(entry);
-      }
-    }
-    flushCards();
-    return rows;
-  }, [entries, columnCount]);
-
-  const rowVirtualizer = useVirtualizer({
-    count: layoutRows.length,
-    getScrollElement: () => parentRef.current,
-    estimateSize: (i) => (layoutRows[i]?.kind === 'section' ? SECTION_HEIGHT : CARD_HEIGHT + GAP),
-    overscan: 3,
-  });
-
-  const virtualItems = rowVirtualizer.getVirtualItems();
-  const totalSize = rowVirtualizer.getTotalSize();
-
-  const gridStyle = useMemo(
-    () => ({
-      gridTemplateColumns: `repeat(${columnCount}, minmax(0, 1fr))`,
-      gap: `${GAP}px`,
-    }),
-    [columnCount],
-  );
-
-  return (
-    <div ref={parentRef} className="h-full overflow-auto p-3">
-      <div style={{ height: totalSize, position: 'relative' }}>
-        {virtualItems.map((vRow) => {
-          const row = layoutRows[vRow.index];
-          if (!row) return null;
-
-          if (row.kind === 'section') {
-            return (
-              <div
-                key={vRow.key}
-                className="absolute left-0 right-0 flex items-end"
-                style={{
-                  height: SECTION_HEIGHT,
-                  transform: `translateY(${vRow.start}px)`,
-                }}
-              >
-                <div className="flex w-full items-center gap-3 pb-1">
-                  <span
-                    className="text-xs tracking-wide text-muted-foreground"
-                    style={{ fontFamily: 'var(--font-display)', fontWeight: 700 }}
-                  >
-                    {row.label}
-                  </span>
-                  <div className="h-px flex-1 bg-border" />
-                </div>
-              </div>
-            );
-          }
-
-          return (
-            <div
-              key={vRow.key}
-              className="absolute left-0 right-0 grid"
-              style={{
-                ...gridStyle,
-                transform: `translateY(${vRow.start}px)`,
-              }}
-            >
-              {row.items.map((entry) =>
-                entry.kind === 'ap' ? (
-                  <ApCard key={`ap-${entry.group.subcategory}`} group={entry.group} onClick={() => onSelect(entry)} />
-                ) : entry.kind === 'book' ? (
-                  <BookCard
-                    key={entry.book.id}
-                    book={entry.book}
-                    onClick={() => onSelect(entry)}
-                    onContextMenu={onBookContextMenu}
-                  />
-                ) : null,
-              )}
-            </div>
-          );
-        })}
-      </div>
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Cards
-// ---------------------------------------------------------------------------
-
-function BookCard({
-  book,
-  onClick,
-  onContextMenu,
-}: {
-  book: Book;
-  onClick: () => void;
-  onContextMenu?: (e: React.MouseEvent, book: Book) => void;
-}) {
-  const [coverError, setCoverError] = useState(false);
-  const [coverUrl, setCoverUrl] = useState<string | null>(null);
-
-  useEffect(() => {
-    setCoverError(false);
-    setCoverUrl(null);
-    if (book.ingested) {
-      api
-        .booksGetCoverUrl(book.id)
-        .then(setCoverUrl)
-        .catch((err) => {
-          console.error(`Failed to load cover for book ${book.id}:`, err);
-          setCoverError(true);
-        });
-    }
-  }, [book.id, book.ingested]);
-
-  return (
-    <button
-      type="button"
-      onClick={onClick}
-      onContextMenu={onContextMenu ? (e) => onContextMenu(e, book) : undefined}
-      className="group relative flex flex-col overflow-hidden rounded-md border border-border bg-card text-left transition-all hover:border-primary/60"
-      style={{ height: CARD_HEIGHT }}
-      title={effectiveTitle(book)}
-    >
-      <CoverArea
-        coverUrl={coverUrl}
-        coverError={coverError}
-        onCoverError={() => setCoverError(true)}
-        ingested={book.ingested}
-        title={effectiveTitle(book)}
-      />
-      {book.aiSystem && book.aiSystem !== 'PF2e' && <SystemBadge system={book.aiSystem} />}
-      {book.ruleset && <RulesetBadge ruleset={book.ruleset} />}
-      <HoverMeta title={effectiveTitle(book)} pageCount={book.pageCount} />
-    </button>
-  );
-}
-
-function ApCard({ group, onClick }: { group: ApGroup; onClick: () => void }) {
-  const coverBook = group.parts[0]?.book;
-  const [coverError, setCoverError] = useState(false);
-  const [coverUrl, setCoverUrl] = useState<string | null>(null);
-
-  useEffect(() => {
-    setCoverError(false);
-    setCoverUrl(null);
-    if (coverBook?.ingested) {
-      api
-        .booksGetCoverUrl(coverBook.id)
-        .then(setCoverUrl)
-        .catch((err) => {
-          console.error(`Failed to load cover for book ${coverBook.id}:`, err);
-          setCoverError(true);
-        });
-    }
-  }, [coverBook?.id, coverBook?.ingested]);
-
-  const totalPages = apTotalPages(group);
-
-  return (
-    <button
-      type="button"
-      onClick={onClick}
-      className="group relative flex flex-col overflow-hidden rounded-md border border-border bg-card text-left transition-all hover:border-primary/60"
-      style={{ height: CARD_HEIGHT }}
-      title={`${group.subcategory} (${group.parts.length}-part Adventure Path)`}
-    >
-      <CoverArea
-        coverUrl={coverUrl}
-        coverError={coverError}
-        onCoverError={() => setCoverError(true)}
-        ingested={coverBook?.ingested ?? false}
-        title={group.subcategory}
-      />
-      {/* AP badge */}
-      <div className="pointer-events-none absolute right-1 top-1 flex items-center gap-0.5 rounded bg-primary/90 px-1 py-0.5 text-[9px] font-semibold text-primary-foreground shadow-xs">
-        <Layers className="h-2.5 w-2.5" />
-        {group.parts.length}
-      </div>
-      <HoverMeta
-        title={group.subcategory}
-        subtitle={`${group.parts.length}-part AP${totalPages != null ? ` · ${totalPages} pages` : ''}`}
-      />
-    </button>
-  );
-}
-
-// Shared cover image area — fills the entire card.
-function CoverArea({
-  coverUrl,
-  coverError,
-  onCoverError,
-  ingested,
-  title,
-}: {
-  coverUrl: string | null;
-  coverError: boolean;
-  onCoverError: () => void;
-  ingested: boolean;
-  title: string;
-}) {
-  return (
-    <div className="absolute inset-0 overflow-hidden bg-muted">
-      {coverUrl && !coverError ? (
-        <img
-          src={coverUrl}
-          alt={title}
-          loading="lazy"
-          onError={onCoverError}
-          style={{
-            width: '100%',
-            height: '100%',
-            objectFit: 'cover',
-            objectPosition: 'top',
-            display: 'block',
-          }}
-          className="transition-transform group-hover:scale-[1.03]"
-        />
-      ) : (
-        <div className="flex h-full w-full flex-col items-center justify-center gap-1 p-2 text-center">
-          <Library className="h-6 w-6 text-muted-foreground/40" />
-          <span className="text-[9px] leading-tight text-muted-foreground/60">
-            {ingested ? 'Cover unavailable' : 'Not yet opened'}
-          </span>
-        </div>
-      )}
-    </div>
-  );
-}
-
-// Metadata overlay shown on hover at the bottom of the card.
-function HoverMeta({ title, pageCount, subtitle }: { title: string; pageCount?: number | null; subtitle?: string }) {
-  return (
-    <div className="pointer-events-none absolute inset-x-0 bottom-0 translate-y-full bg-black/75 px-2 py-1.5 backdrop-blur-sm transition-transform group-hover:translate-y-0">
-      <div className="truncate text-xs font-medium leading-tight text-white">{title}</div>
-      {subtitle && <div className="text-[10px] text-white/70">{subtitle}</div>}
-      {!subtitle && pageCount != null && <div className="text-[10px] text-white/70">{pageCount} pages</div>}
-    </div>
-  );
-}
-
-function SystemBadge({ system }: { system: string }) {
-  return (
-    <div className="pointer-events-none absolute left-1 top-1 rounded bg-amber-600/90 px-1 py-0.5 text-[9px] font-semibold uppercase text-white shadow-xs">
-      {system}
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Right-click context menu for moving books between categories / systems
-// ---------------------------------------------------------------------------
-
-const CTX_CATEGORIES = ['Rulebook', 'Adventure Path', 'Adventure', 'Setting', 'Supplement'] as const;
-const CTX_SYSTEMS = ['PF2e', '5e', 'Generic'] as const;
-
-function BookContextMenu({
-  book,
-  x,
-  y,
-  onClose,
-  onUpdateMeta,
-}: {
-  book: Book;
-  x: number;
-  y: number;
-  onClose: () => void;
-  onUpdateMeta: (bookId: number, fields: { aiSystem?: string; aiCategory?: string }) => void;
-}) {
-  const ref = useRef<HTMLDivElement>(null);
-
-  // Close on click-outside or Escape.
-  useEffect(() => {
-    const handleClick = (e: MouseEvent) => {
-      if (ref.current && !ref.current.contains(e.target as Node)) onClose();
-    };
-    const handleKey = (e: KeyboardEvent) => {
-      if (e.key === 'Escape') onClose();
-    };
-    document.addEventListener('mousedown', handleClick);
-    document.addEventListener('keydown', handleKey);
-    return () => {
-      document.removeEventListener('mousedown', handleClick);
-      document.removeEventListener('keydown', handleKey);
-    };
-  }, [onClose]);
-
-  // Clamp position so menu doesn't overflow viewport.
-  const style = useMemo(() => {
-    const menuW = 180;
-    const menuH = 280;
-    const left = Math.min(x, window.innerWidth - menuW - 8);
-    const top = Math.min(y, window.innerHeight - menuH - 8);
-    return { position: 'fixed' as const, left, top, zIndex: 9999 };
-  }, [x, y]);
-
-  const curCat = effectiveCategory(book);
-  const curSys = effectiveSystem(book);
-
-  return (
-    <div
-      ref={ref}
-      style={{ ...style, backgroundColor: 'hsl(var(--popover))' }}
-      className="min-w-[160px] rounded-md border border-border py-1 shadow-lg"
-    >
-      <div className="px-2 py-1 text-[10px] font-semibold uppercase tracking-wider text-muted-foreground">Category</div>
-      {CTX_CATEGORIES.map((cat) => (
-        <button
-          key={cat}
-          type="button"
-          className={cn(
-            'flex w-full items-center gap-2 px-2 py-1 text-left text-xs transition-colors hover:bg-accent',
-            curCat === cat && 'text-foreground font-medium',
-            curCat !== cat && 'text-muted-foreground',
-          )}
-          onClick={() => {
-            onUpdateMeta(book.id, { aiCategory: cat });
-            onClose();
-          }}
-        >
-          <Check className={cn('h-3 w-3', curCat === cat ? 'opacity-100' : 'opacity-0')} />
-          {cat}
-        </button>
-      ))}
-      <div className="mx-2 my-1 h-px bg-border" />
-      <div className="px-2 py-1 text-[10px] font-semibold uppercase tracking-wider text-muted-foreground">System</div>
-      {CTX_SYSTEMS.map((sys) => (
-        <button
-          key={sys}
-          type="button"
-          className={cn(
-            'flex w-full items-center gap-2 px-2 py-1 text-left text-xs transition-colors hover:bg-accent',
-            curSys === sys && 'text-foreground font-medium',
-            curSys !== sys && 'text-muted-foreground',
-          )}
-          onClick={() => {
-            onUpdateMeta(book.id, { aiSystem: sys });
-            onClose();
-          }}
-        >
-          <Check className={cn('h-3 w-3', curSys === sys ? 'opacity-100' : 'opacity-0')} />
-          {sys}
-        </button>
-      ))}
-    </div>
-  );
-}
-
-function RulesetBadge({ ruleset }: { ruleset: 'legacy' | 'remastered' }) {
-  return (
-    <div
-      className={cn(
-        'pointer-events-none absolute right-1 top-1 rounded px-1 py-0.5 text-[9px] font-semibold uppercase shadow-xs',
-        ruleset === 'remastered' ? 'bg-primary/90 text-primary-foreground' : 'bg-muted-foreground/80 text-background',
-      )}
-    >
-      {ruleset === 'remastered' ? 'R' : 'L'}
     </div>
   );
 }
